refactored media from transmissionlayer created media in CommunicationService created RealAudio media refactored PaWrapper to root

+++ CommunicationService/TransmissionLayer/AcousticTL/helperFunctions.h
@@ -13,8 +13,8 @@ inline std::vector<std::pair<int, float>> goertzelFilter(const std::vector<float
     std::vector<std::pair<int, float>> returnAmpFreq;
 
     for (int i = 0; i < freqs.size(); ++i) {
-        double k = 0.5 + ((blockSize*freqs[i])/(sampleRate));
-        double w = (2 * M_PI / blockSize) * k;
+        double k = 0.5 + blockSize * freqs[i] / sampleRate;
+        double w = 2 * (double)M_PI / blockSize * k;
         double cosine = cos(w);
         double sine = sin(w);
         double coeff = 2 * cosine;
